requirements.txt
streamlit>=1.35.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
